fix(bridge): move WakeLockClient to api layer per Aristotle review
