+++ bridge/app/lib/src/api/linux_wake_lock_api.dart
@@ -1,7 +1,7 @@
 import "dart:async";
 import "dart:io";
 
-import "package:sesori_bridge/src/foundation/wake_lock_client.dart";
+import "package:sesori_bridge/src/api/wake_lock_client.dart";
 import "package:sesori_plugin_interface/sesori_plugin_interface.dart" show Log;
 
 typedef ProcessStarter = Future<Process> Function(
