+++ bridge/app/lib/src/api/wake_lock_client.dart
